requirements.txt
aiohttp==3.7.3
alembic==1.5.4
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.6.0
gitdb==4.0.5
GitPython==3.1.12
idna==3.1
loguru==0.5.3
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.3.23
typing-extensions==3.7.4.3
yarl==1.6.3
